Modifications
Modifications to the Library, LibraryBook and LibraryBookGeneric
classes.

For Library I implemented the lookup, checkin and checkout methods.

For LibraryBook I implemented getHolder, getDueDates, checkin, and
checkout methods

For the LibraryBookGeneric I modified the String type for the library
book's holder in my original class to Type in the new class.

*More detail listen in the comments